Some clean up for TFTP and HTTP iPXE binary serving:
